Create component md
